refactor: introduced commander to bin
