+++ src/lib/interfaces/SpriteImage.ts
