+++ tsconfig.json
@@ -9,5 +9,6 @@
     "strict": true,
     "skipLibCheck": true
   },
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist"],
+  "include": ["src/bin/index.ts"]
 }
